+++ project_1/src/ServerBase.cpp
@@ -3,6 +3,6 @@
 ServerBase::ServerBase(std::uint64_t id)
     : id(id) { }
 
-std::uint32_t ServerBase::GetID() const {
+std::uint64_t ServerBase::GetID() const {
     return id;
 }
